pyproject.toml
[tool.poetry]
name = "clippinator"
version = "0.1.0"
description = ""
authors = ["Lev <ennucore@gmail.com>", "Timofey fedoseev.timofey@icloud.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
langchain = "^0.0.235"
rich = "^13.3.4"
typer = "^0.7.0"
python-dotenv = "^1.0.0"
google-search-results = "^2.4.2"
wolframalpha = "^5.0.0"
python-telegram-bot = "^20.2"
unidiff = "^0.7.5"
patch = "^1.16"
tiktoken = "^0.3.3"
html2text = "^2020.1.16"
beautifulsoup4 = "^4.12.2"
selenium = "^4.9.1"
setuptools = "^67.8.0"
inquirer = "^3.1.3"
anthropic = "^0.3.4"


[tool.poetry.group.dev.dependencies]
ipython = { version = "^8.12.0", optional = true }
bandit = "^1.7.5"
pylint = { version = "^2.17.2", optional = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
clippinator = "clippinator.__main__:app"
